apps/backend/pyproject.toml
[project]
name = "smart-data-platform"
version = "1.0.0"
description = "Enterprise-grade intelligent data management platform"
authors = [
    {name = "One Data Studio", email = "info@onedatastudio.io"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    # Web Framework
    "fastapi>=0.109.2",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",

    # Database
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",

    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",

    # Cache
    "redis>=5.0.1",
    "hiredis>=2.3.2",

    # Task Queue
    "celery>=5.3.4",
    "flower>=2.0.1",

    # Storage
    "minio>=7.2.0",

    # ML/AI
    "mlflow>=2.9.0",
    "transformers>=4.37.0",
    "torch>=2.2.0",
    "sentence-transformers>=2.3.0",
    "chromadb>=0.4.18",

    # LLM
    "openai>=1.10.0",
    "langchain>=0.1.0",
    "httpx>=0.27.0",

    # Utilities
    "python-dateutil>=2.8.2",
    "pytz>=2024.1",
    "tenacity>=8.2.3",
    "pydantic[email]>=2.6.0",

    # Monitoring
    "prometheus-client>=0.20.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.4",
    "pytest-mock>=3.12.0",
    "httpx>=0.27.0",  # for TestClient
    "black>=24.3.0",
    "isort>=5.13.2",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "ipdb>=0.13.13",
    "ipython>=8.22.0",
]

[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
